requirements.txt
torch
transformers
langchain
sentence-transformers
datasets
langchain-community
faiss-gpu
faiss-cpu
pandas
numpy
openpyxl
torchvision
timm
einops
ftfy
pytrec_eval
llama-cpp-python
xlrd
matplotlib
seaborn
lime
shap
nltk
spacy
gensim
wordcloud
langdetect
textblob
transformers==4.48.0
rouge
contractions
tokenizers
tensorflow
scipy
lbl2vec
imblearn
tf-keras